Only check sql_require_primary_key for InnoDB/MyRocks engine
Summary: In production we may have blackhole tables that doesn't have PK and we would end up not being able to create those tables during copy. The fix is to align with 5.6 behavior and only check this for MyRocks/InnoDB tables and ignore other engines that don't store user data.

Differential Revision: D23669893

fbshipit-source-id: 30e4ac20412