Making rocksdb_max_bottom_pri_background_compactions dynamic
Summary:
This diff makes rocksdb_max_bottom_pri_background_compactions
dynamic, so that it can be increased or decreased number of threads
online. Changing from non zero to zero is not supported, due to RocksDB
limitations. Changing the number of threads may take time, so it is done
outside of global mutex.
This diff also changed the variable description for better
understanding.

This diff was ported from
https://github.com/facebook/mysql-5.6/commit/49b047091d9fcc49f92abb11e3e80d192a5f6efb
in 5.6.

Differential Revision: D23799402

fbshipit-source-id: 585ae917a36