fix comparison semantics of legacy json_extract
Summary:
5.6 json_extract returns a string type and this string type is then used to compare
against other operands.
8.0 version of json_extract however return a json type. If any operand is compared
against json_extract, the operand itself is converted to json type, eventually calling into
Json_wrapper::compare. We would want to keep this behavior as 8.0 json_extract is
the way forward.

In case any of the operands is a string, it is not safe to use Json_wrapper::compare since
it stores string differently (without quotes). Also Json_wrapper::compare does not
compare string with any other types. However 5.6 implementation did.

This change implements a custom comparison logic for json_extract. This should
hopefully fix all the divergent comparison output we are seeing in 8.0 compared to
5.6

Differential Revision: D23946125

fbshipit-source-id: 6bdfb280c4e