Fix direct_io check
Summary:
In wsenv, FileSystem::ReopenWritableFile isn't supported, so if use_direct_reads = on, we would simply fail to bootstrap the database, with an empty error message.

In the case of testing for direct_io, reopening the DIRECT_CHECK or truncating the file makes no difference, so this simply switches to NewWritableFile which truncates the file and then opens it for write.

Also use ToString for a better error reporting, instead of getState (which returns empty string in this case). ToString would include getState anyway.

Differential Revision: D23952964

fbshipit-source-id: 217cb2d9895