Port: reinitializing previous_gtid_set_map in binlog_change_to_apply
Summary:
Ported D23801889
**Summary**
A recent crash was discovered in SHOW BINARY LOGS WITH GTID
right after a raft bootstrap. The crash was because the previous_gtid_set_map
map is not uninitialized properly during binlog_change_to_apply. This
was an overlook. binlog_change_to_apply is unique in the repointing
because the apply logs previous gtid set is initialized with the
gtid_executed of the previous primary or from the previous applier. Its
not initialized from files. When raft is bootstrapped for first time there could
be old apply-3306.index file on disk and several apply-3306.# files.
However if an instance bootstraps as FOLLOWER, it executes
binlog_change_to_apply which has an old previous_gtid_set_map containing
mappings from binary-3306.# files to its GTID set. Now if we do
SHOW BINARY LOGS WITH GTID a predictable crash happens.
As a part of this fix, we clean up/purge apply-3306.index except the
last active file. After this a trimmed version of init_gtid_sets
is called, which sets previous_gtid_set_map properly.

Reviewed By: luqun

Differential Revision: D24631274

fbshipit-source-id: c3611cb7326