Handle proper exit handling for raft listener thread
Summary:
Raft listener threads were predictably getting stuck during shutdown, because close_connections would not be able
to kill it.

Making the init and deinit of raft listener thread tied to the registration of raft observer in the server. This is intuitive
because it is the point where the plugin and server first communicate.

On the shutdown, path added another hook so that the shutdown happens before close_connections

Porting changes:
1. FOREACH_OBSERVER in 8.0 remove thd parameter, thus doesn't port related change

Differential Revision: D24789826

fbshipit-source-id: 32bf04f47f1