Fix incorrect socket path length check
Summary:
UNIX domain socket has a max length of 108 (107 character + NULL). mysql-test-path has check_socket_path_length that checks the length and make sure the socket path fits. It does that twice: first time for the temp path (/data/users/yzha/mysql/mysql80-fork-this-is-a-longdir/_build-8.0-Debug/mysql-test/var/tmp) and then the 2nd time for individual worker (/data/users/yzha/mysql/mysql80-fork-this-is-a-longdir/_build-8.0-Debug/mysql-test/var/tmp/12). If the first check fails, it will fall back to use tmp path (/tmp/) to workaround the limitation, but not the second check, since it assumes the first check would catch it.

However, the 2nd check already has the child number in the path (/data/users/yzha/mysql/mysql80-fork-this-is-a-longdir/_build-8.0-Debug/mysql-test/var/tmp/12) but it counts the opt_parallel again so the childnum component is double counted. As a result, it is possible to succeed at the first check but then fail at the second check due to the check counting, so the worker would simply fail.

To avoid double counting, we just pass 0 as parallel to check_socket_path_length in the 2nd check.

Differential Revision: D25152364

fbshipit-source-id: 50f7ff2e167