Wrapping last semi-sync acked pos in std::atomic to avoid locking in some scenarios
Summary:
We'll now lock the mutex only if we need to wait for last acked
pos to update. Otherwise, we just check the current dump thread pos
against the last acked pos atomically and if current is less that last
acked we sent the event without locking the mutex.

Differential Revision: D25772521

Squash with D16267709

fbshipit-source-id: f710bea8667