Properly fallback if executing unsupported case
Summary:
In most cases bypass parser `select_parser` should identify unsupported case and fallback to regular query. However, there are a few checks that are expensive and is done as part of execution, and we should properly fallback in those case as well. In a future version maybe we should refactor it so that those are another phase of parsing as well, which would also help bypass RPC.

Reference Patch: https://github.com/facebook/mysql-5.6/commit/5452df8218acb45b7664a8eeb4542254a230730c

Reviewed By: luqun

Differential Revision: D26564781

fbshipit-source-id: 07297325810