Skip decoding completely during count and improve SK count perf by 20x
Summary:
8.0 is taking a new approach with doing counts - MySQL will call handler::records / records_with_index which will then use rnd_init+rnd_next / index_init+index_next to iterate through everything. If you try count(*) you'll see MyRocks 8.0 is actually much slower when iterating the table with secondary keys than 5.6 (the optimizer picks secondary key for whatever reason).

This is because MySQL no longer calls `handler::extra(KEY_READ)` so we no longer get the hints to only decode the key - instead we'll see that the SK is not covering so we fall back to PK, which dramatically slows down by *a lot*.

This fixes it by adding a m_iteration_only = true hint -  with this hint we'll just quickly iterate through everything without any decoding nor falling back to PK.

Note even with this fix, if you compare MyRocks 8.0 with InnoDB 8.0, you'll see that InnoDB 8.0 is also much faster - this is because InnoDB overrides the implementation and iterate the records directly in parallel (default to 4). This is something we can do as well.

Reviewed By: Pushapgl

Differential Revision: D26272733

fbshipit-source-id: 53b1ae328f6