Add server_cpu and warnings to query response attributes
Summary:
This diff adds server CPU time to query response attributes. The key of the attribute is "server_cpu" and the value is the stringified version of the cpu time.

This diff also adds warnings information to query response attributes. The key of the attribute is "warnings" and the value is the string that serializes the warnings information. Since there could be more than one warnings for a statement, the value is a list of pairs included with in brackets separated by commas. The first value of the pair is the error number (code) and the second value of the pair is the message text and these are separated by a comma.

The following example shows the query response attribute for 'warnings' where two warnings are raised for the statement:
**1265,Data truncated for column 'c1' at row 1),(1264,Out of range value for column 'c2' at row 1)**

The variable 'response_attrs_contain_warnings' controls this feature and is disabled by default.

If the query raises any errors then the response OK packet is not set to the client. So in this case putting warnings into response attributes is skipped.

Reviewed By: george-reynya

Differential Revision: D26598990

fbshipit-source-id: e3691a0c366