Add locking when reading index cardinality in InnoDB
Summary:
When updating index cardinality in InnoDB, the stats are first zero'd out via `dict_stats_empty_index` and then populated. Although there is locking on the write path, when updating index stats, it looks like we usually do not have locking on the read path in innodb. This can lead to bad plans, as we would return bad index stats.

The fix is to take a shared lock on the read paths.

https://bugs.mysql.com/bug.php?id=103210

Reviewed By: luqun, yizhang82

Differential Revision: D27545419

fbshipit-source-id: 92de08d9197