Add retry when opening doulewrite buffer
Summary:
This fixes the `main.log_buffered-big` test.

The test intentionally tries to start another mysqld instance on the same data files while an existing instance still has those files open. The expected behaviour is that the second instance blocks on opening the system tablespace ibdata1 file until the first instance shuts down.

However with 8.0.20, the doublewrite buffer has been separated out to their own files and because we open and close the ibdata1 file and the doublewrite files in the same order, it's possible for the first instance to have closed the ibdata1 file but not the doublewrite file, while the second instance has opened the ibdata1 file, but fails to open the doublewrite files.

The fix to also add retry to the doublewrite file opening logic. The alternative is to reorder the file open/closes but that seemed riskier.

https://bugs.mysql.com/bug.php?id=103338

Reviewed By: hermanlee

Differential Revision: D27797914

fbshipit-source-id: 9400e329101