Replication lag timestamp occassionally gets updated to now()
Summary:
When a secondary is lagging and the relay log space is full, the I/O
thread may start/stop fetching more logs from the primary. There are
certain conditions where the SQL thread enters the codepath in
coord_handle_partial_binlogged_transaction(). This likely happens if the
I/O thread is disconnected in the middle of fetching a transaction
group. The next time the I/O thread reconnects, it fetches a format
descriptor event. When the SQL thread finds a format descriptor event in
the middle of a group, it generates two query log events on demand to
rollback the transaction.

However, these events have their timestamps set to now(). When the
replication workers process these events, they update
last_master_timestamp to the latest value. Since last_master_timestamp
is always monotonically increasing, the SBM reported by the secondary
drops down to 0.

Fix this by setting the timestamps on these query log events to
the current last_master_timestamp.

Reviewed By: abhinav04sharma

Differential Revision: D27975963

fbshipit-source-id: 086ec2f2a56