Add variable to prevent serializing legacy json types
Summary:
JSON objects produced by the FB version of JSON_EXTRACT compares differently than objects produced by the upstream version. This was done to maintain 5.6 compatibility. The way this was implemented was by recording this fact in `Json_wrapper::m_legacy_json`.

However, this fact is not persisted, which means that if the json object is serialized, we will compare assuming this was an upstream json object, resulting in incorrect results.

The easiest way around this is to ask clients to rewrite their query (possibly by casting the json object into a string). In order to surface these issues, attempts to serialize FB json objects will either produce a warning or an error.

Reviewed By: atish2196

Differential Revision: D28551653

fbshipit-source-id: 49914e83b3f