SHOW SLAVE STATUS shows real error
Summary: In parallel replication SHOW SLAVE STATUS doesn't show the actual error. It actually creates the real error (Slave_reporting_capability::va_report) and set m_last_error right after setting coordinator error so we just need to reverse the order of the two and show the real error in coordinator. Accessing the last error is protected by err_lock and also fallback to previous message in case the error number is 0 and can't be trusted, just in case.

Reviewed By: hermanlee, Pushapgl

Differential Revision: D28925760

fbshipit-source-id: f351b8ca3e6