Avoid considering skip scan keys in queries with large IN lists
Summary:
Skip scan considers all keys containing any fields found in the WHERE clause, meaning that range analysis will examine a lot more keys due to skip scan. This can cause it to run out memory, and fall back to full table scan. To help mitigate this, do not consider skip scan keys if we find large IN lists. Note that this isn't a complete solution because queries can also generate a lot of ranges through other mechanisms too.

This limit is controlled by the variable `optimizer_skip_scan_in_list_limit`. and a warning is produced if we avoided skip scan keys due to the IN list limit.

Note that it is still possible that a skip scan plan is produced from the pre-existing range scan keys. This is not a problem, since we would have needed to build range trees for range plans anyway.

Reviewed By: satya-valluri

Differential Revision: D30428092

fbshipit-source-id: 37f62aeae4a