Make FB added client errors consistent across versions
Summary: We have run into multiple problems when upgrading the MySQL client library in fbcode when the FB-added client error messages change their value because we have to adjust them when Oracle adds more client errors.  To try to avoid this continuing to happen in the future, add a bunch of place holder client error messages to push the FB-added ones way down the list.

Reviewed By: aditya-jalan, lth

Differential Revision: D31160326

Squash with D6325225

fbshipit-source-id: 3c5e4d064e2