Extend command api to support OK Metadata
Summary:
According to oracle the session trackers are considered
a classic protocol detail. We rely on the fields of
session trackers being returned to track state and
return metadata. Add a metadata ptr to the handle_ok
callback.

Consumers will always be responsible for checking the
ptr value isn't null. Classic protocol will ignore the field
Plugin protocols can consume it as needed, but by default will
not since it's our feature.

Reviewed By: fadimounir

Differential Revision: D31315069

fbshipit-source-id: 2ed94da02e8