Track savepoints in session state changes
Summary:
We can't detach when there are savepoints. They're normally supposed to
be in transactions, but with autocommit = 0 you can create a savepoint
outside of a transaction.

We discussed about setting the in_transaction for this but that has too
many other effects and is risky. Instead, set the session state flag
which is consumed by the clients to tell them it's not save to switch
connections.

Reviewed By: abal147

Differential Revision: D31664148

fbshipit-source-id: c3d89760ae0