Improve rpl.rpl_binlog_wait_for_hlc for valgrind testing
Summary:
This test fails in valgrind because the hlc timeouts were too
low. Increase the timeouts from 1000ms to 10000ms to give valgrind
run more time for replication log events to be processed by the
secondaries.

Also add WHERE conditions for the SELECT * queries to filter out
unwanted rows. While HLC guarantees a lower bound, it does not guarantee
an upper bound.  It is possible for a secondary to process binlog events
past the desired HLC, leading to SELECT * returning more rows than was
recorded in the results file.

Squash with D24263597

Reviewed By: Pushapgl

Differential Revision: D31848160

fbshipit-source-id: 66e42255f42