FB8-265: main.innodb_pk_extension_on fail in 8.0.23 after porting from 8.0.20
https://jira.percona.com/browse/FB8-265

Post push fix.

After fixing PB8-265 (commit 865b375a), rocksdb_cf_per_partition MTR
test started to fail in the following way:
-1	SIMPLE	t2	custom_p2	ref	col3	col3	258	const	1	50.00	Using index condition; Using where
+1	SIMPLE	t2	custom_p2	ref	col3	col3	258	const	1	50.00	Using index condition

Cause:
Fixing PB8-265 causes proper recognition that PK parts are added to SK.
That causes 'Using where' not to be used as all condition fields are
part of SK (col2 is not in explicit way, but it is part of PK which was
added to SK)

Solution:
1. Existing test case re-recorded
2. Added test case which uses field not being part of a key, which
causes 'Using where' strategy appearing in 'Extra' column.